requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.1.7
django-admin-honeypot==1.1.0
django-browser-reload==1.7.0
django-cors-headers==3.14.0
django-filter==22.1
django-grappelli==3.0.4
django-heroku==0.3.1
django-tailwind==3.5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.0.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
